Move hint placement code to separate file
